% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_metadata.R
\name{acquire_metadata}
\alias{acquire_metadata}
\title{Download and extract metadata}
\usage{
acquire_metadata(method = "curl")
}
\arguments{
\item{method}{method passed to /code{download.file}}
}
\description{
/code{acquire_metadata} downloads (using curl as a default)
to download the metadata from Project Gutenberg. The compressed
file is stored in /code{tempdir()} and extracted there using
/code{untar}.
}
