% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_metadata.R
\name{cache_metadata}
\alias{cache_metadata}
\title{Acquire and cache metadata}
\usage{
cache_metadata(lib = "~/gutenbergr")
}
\arguments{
\item{lib}{The location to store the cached metadata}
}
\description{
/code{cache_metadata} calls /code{acquire_metadata}
to download and temporarily store the Project Gutenberg
metadata. It then formats and caches the metadata so
that it can be used again. Downloading and extracting
takes a long time, so it is useful to cache the data
}
