% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rdf_metadata.R
\name{check_metadata_file}
\alias{check_metadata_file}
\title{Check if file is in the expected metadata format}
\usage{
check_metadata_file(xml_file, root_ref = "pgterms:ebook")
}
\description{
/code{check_metadata_file} takes an XML file and ensures that it is
of the expected format for a Project Gutenberg metadata file for a
single work.
}
