% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_download.R
\name{check_wget}
\alias{check_wget}
\title{Check is \code{wget} is installed}
\usage{
check_wget()
}
\value{
Boolean. Is \code{wget} installed?
}
\description{
Check is \code{wget} is installed
}
\keyword{internal}
