% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cycle_through_urls}
\alias{cycle_through_urls}
\title{Loop through potential file URLs}
\usage{
cycle_through_urls(url, ext = c(".zip", ".txt"))
}
\arguments{
\item{url}{URL to a file}

\item{ext}{Extension of the file to read}
}
\description{
Loop through potential file URLs
}
