% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discard_start_while}
\alias{discard_start_while}
\title{Discard all values at the start of .x while .p is true}
\usage{
discard_start_while(.x, .p)
}
\arguments{
\item{.x}{Vector}

\item{.p}{Logical vector}
}
\value{
A vector the same type as \code{.x} with leading elements removed.
}
\description{
Discard all values at the start of .x while .p is true
}
\keyword{internal}
