% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dl_and_read}
\alias{dl_and_read}
\title{Download and read a file}
\usage{
dl_and_read(url, tmp)
}
\arguments{
\item{url}{URL to a file}

\item{tmp}{Path to a temporary file.}
}
\value{
A character vector with one element for each line.
}
\description{
Download and read a file
}
\keyword{internal}
