% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_works.R
\name{download_by_id}
\alias{download_by_id}
\title{Download gutenberg work by id and format}
\usage{
download_by_id(id, output.dir = NULL, sleep_time = 5,
  file_format = "text/plain", zip = FALSE)
}
\arguments{
\item{id}{a numeric vector of Project Gutenberg work ids}

\item{output.dir}{}

\item{sleep_time}{time to sleep between file downloads, in seconds}

\item{file_format}{format of file to download}

\item{zip}{download a zip file instead of raw if available}
}
\description{
\code{download_by_id} performs the actual downloading  of a work's text
from Project Gutenberg. Given an id and file format, the file is downloaded
and either stored in an output directory or returned to the console.
}
