% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_works.R
\name{download_gutenberg}
\alias{download_gutenberg}
\title{Download gutenberg work by id, title pattern, author, or title}
\usage{
download_gutenberg(gutenberg_id, pattern, authors, titles,
  cache_path = getOption("gutenbergr_cache"), ...)
}
\arguments{
\item{gutenberg_id}{numeric vector of Project Gutenberg work ids}

\item{pattern}{named character vector of search patterns to download}

\item{authors}{character vector of authors to download}

\item{titles}{character vector of titles to download}

\item{cache_path}{location of cached metadata}

\item{...}{other parameters passed to \code{download_by_id}}
}
\description{
Download gutenberg work by id, title pattern, author, or title
}
\examples{

}
