% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_download.R
\name{flatten_gutenberg_id}
\alias{flatten_gutenberg_id}
\title{Subset gutenberg_id from df if necessary}
\usage{
flatten_gutenberg_id(gutenberg_id)
}
\arguments{
\item{gutenberg_id}{A vector of Project Gutenberg IDs, or a data frame
containing a \code{gutenberg_id} column, such as from the results of
\code{\link[=gutenberg_works]{gutenberg_works()}}.}
}
\value{
A character vector of gutenberg_ids.
}
\description{
Subset gutenberg_id from df if necessary
}
\keyword{internal}
