% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_works.R
\name{get_combined_metadata}
\alias{get_combined_metadata}
\title{Combine metadata for works and authors}
\usage{
get_combined_metadata(cache_path = getOption("gutenbergr_cache"))
}
\arguments{
\item{cache_path}{location of cached metadata.}
}
\description{
\code{get_combined_metadata} checks that the metadata files
for author and work are loaded; if not, their cache files
are loaded from \code{cache_path}. The two metadata tables
are joined to get complete metadata about each work.
}
