% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_works.R
\name{get_gutenberg_paths}
\alias{get_gutenberg_paths}
\title{Get download path to gutenbergr work}
\usage{
get_gutenberg_paths(id, file_format = "text/plain", zip = TRUE,
  cache_path = getOption("gutenbergr_cache"))
}
\arguments{
\item{id}{a numeric vector of Project Gutenberg work ids}

\item{file_format}{file format to download}

\item{zip}{download zipped format instead of raw if available}

\item{cache_path}{location of cached metadata}
}
\description{
\code{get_gutenberg_paths} uses the loaded or cached metadata
about the formats of Gutenberg works to determine the url of
the file corresponding to a work and file format.
}
\examples{

get_gutenberg_paths(76)
get_gutenberg_paths(76, zip = FALSE)

get_gutenberg_paths(1:10, "text/html", zip = FALSE)
}
