% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_download.R
\name{gutenberg_add_metadata}
\alias{gutenberg_add_metadata}
\title{Join metadata fields to Gutenberg works}
\usage{
gutenberg_add_metadata(gutenberg_tbl, meta_fields)
}
\arguments{
\item{gutenberg_tbl}{A two column \code{tbl_df} from \code{gutenberg_download}.}

\item{meta_fields}{Additional fields describing each book, such as \code{title}
and \code{author}, to add from \link{gutenberg_metadata}.}
}
\value{
A \code{tbl_df} of the Gutenberg works with joined metadata.
}
\description{
Join metadata fields to Gutenberg works
}
\keyword{internal}
