% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_cache.R
\name{gutenberg_cache_clear_all}
\alias{gutenberg_cache_clear_all}
\title{Clear all files from the Gutenberg cache}
\usage{
gutenberg_cache_clear_all()
}
\value{
The number of files deleted (invisibly).
}
\description{
Deletes all cached \code{.rds} files in the directory currently returned by
\code{\link[=gutenberg_cache_dir]{gutenberg_cache_dir()}}.
}
\examples{
\dontrun{
# Clear entire current cache
gutenberg_cache_clear_all()
}
}
\keyword{cache}
