% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_cache.R
\name{gutenberg_cache_files}
\alias{gutenberg_cache_files}
\title{List cached .rds files}
\usage{
gutenberg_cache_files()
}
\value{
Character vector of full paths
}
\description{
List cached .rds files
}
\keyword{internal}
