% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_cache.R
\name{gutenberg_cache_path}
\alias{gutenberg_cache_path}
\title{Get the active cache directory}
\usage{
gutenberg_cache_path()
}
\value{
Character path to cache dir (may not exist)
}
\description{
Get the active cache directory
}
\keyword{internal}
