% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_cache.R
\name{gutenberg_cache_remove_ids}
\alias{gutenberg_cache_remove_ids}
\title{Delete specific files from the cache}
\usage{
gutenberg_cache_remove_ids(ids, verbose = TRUE)
}
\arguments{
\item{ids}{A numeric or character vector of Gutenberg IDs to remove
from the current cache.}

\item{verbose}{Whether to show the status messages.}
}
\value{
The number of files successfully deleted (invisibly).
}
\description{
Delete specific files from the cache
}
\examples{
\dontrun{
# Remove specific books from cache
gutenberg_cache_remove_ids(c(1, 2))

# Remove silently
gutenberg_cache_remove_ids(1, verbose = FALSE)
}
}
\keyword{cache}
