% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_cache.R
\name{gutenberg_clear_cache}
\alias{gutenberg_clear_cache}
\title{Delete all files in cache path currently returned by \code{\link{gutenberg_cache_path}}}
\usage{
gutenberg_clear_cache()
}
\description{
Delete all files in cache path currently returned by \code{\link{gutenberg_cache_path}}
}
