% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_cache.R
\name{gutenberg_delete_cache}
\alias{gutenberg_delete_cache}
\title{Delete specific files from the cache}
\usage{
gutenberg_delete_cache(ids, verbose = TRUE)
}
\arguments{
\item{ids}{A numeric or character vector of Gutenberg IDs to remove
from the current cache.}

\item{verbose}{Whether to show the status messages.}
}
\value{
The number of files successfully deleted (invisibly).
}
\description{
Delete specific files from the cache
}
\keyword{cache}
