% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_download.R
\name{gutenberg_download}
\alias{gutenberg_download}
\title{Download a book using its Gutenberg ID}
\usage{
gutenberg_download(gutenberg_id, mirror = NULL, strip = TRUE, ...)
}
\arguments{
\item{gutenberg_id}{A vector of Project Gutenberg ID}

\item{mirror}{Optionally a mirror URL to retrieve the books from}

\item{strip}{Whether to strip suspected headers and footers}

\item{...}{Extra arguments passed to \code{\link{gutenberg_strip}}}
}
\description{
Download a book using its Gutenberg ID
}

