% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_download.R
\name{gutenberg_get_mirror}
\alias{gutenberg_get_mirror}
\title{Get the recommended mirror for Gutenberg files}
\usage{
gutenberg_get_mirror()
}
\description{
Get the recommended mirror for Gutenberg files by accessing
the wget harvest path, which is
\url{http://www.gutenberg.org/robot/harvest?filetypes[]=txt}.
Also set the global \code{gutenberg_mirror} option
}

