% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_cache.R
\name{gutenberg_list_cache}
\alias{gutenberg_list_cache}
\title{List files in the cache path currently returned by \code{\link{gutenberg_cache_path}}}
\usage{
gutenberg_list_cache(quiet = FALSE)
}
\arguments{
\item{quiet}{Whether to print the cache directory path}
}
\description{
List files in the cache path currently returned by \code{\link{gutenberg_cache_path}}
}
