% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_download.R
\name{gutenberg_path_from_id}
\alias{gutenberg_path_from_id}
\title{Construct a Project Gutenberg path from an ID}
\usage{
gutenberg_path_from_id(gutenberg_id)
}
\arguments{
\item{gutenberg_id}{A vector of Project Gutenberg IDs, or a data frame
containing a \code{gutenberg_id} column, such as from the results of
\code{\link[=gutenberg_works]{gutenberg_works()}}.}
}
\value{
A character vector of paths.
}
\description{
Construct a Project Gutenberg path from an ID
}
\keyword{internal}
