% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_download.R
\name{gutenberg_tidy_metadata}
\alias{gutenberg_tidy_metadata}
\title{Prepare metadata for a 1-to-1 join}
\usage{
gutenberg_tidy_metadata(metadata, fields, sep = " & ")
}
\arguments{
\item{metadata}{The metadata data frame (usually \code{gutenberg_metadata}).}

\item{fields}{The fields to include in the join.}

\item{sep}{The separator to use for multiple values (e.g., authors).}
}
\value{
A data frame with one row per \code{gutenberg_id}.
}
\description{
Prepare metadata for a 1-to-1 join
}
\keyword{internal}
