% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_download.R
\name{gutenberg_url}
\alias{gutenberg_url}
\title{Construct a Project Gutenberg url}
\usage{
gutenberg_url(gutenberg_id, mirror, verbose)
}
\arguments{
\item{gutenberg_id}{A vector of Project Gutenberg IDs, or a data frame
containing a \code{gutenberg_id} column, such as from the results of
\code{\link[=gutenberg_works]{gutenberg_works()}}.}

\item{mirror}{A mirror URL to retrieve the books from. By default uses the
mirror from \code{\link[=gutenberg_get_mirror]{gutenberg_get_mirror()}}.}

\item{verbose}{Whether to show messages about the Project Gutenberg mirror
that was chosen}
}
\value{
A named character vector of urls
}
\description{
Construct a Project Gutenberg url
}
\keyword{internal}
