% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenbergr-package.R
\docType{package}
\name{gutenbergr-package}
\alias{gutenbergr}
\alias{gutenbergr-package}
\title{gutenbergr: Download and Process Public Domain Works from Project Gutenberg}
\description{
Download and process public domain works in the Project Gutenberg collection \url{http://www.gutenberg.org/}. Includes metadata for all Project Gutenberg works, so that they can be searched and retrieved.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/gutenbergr/}
  \item \url{https://github.com/ropensci/gutenbergr}
  \item Report bugs at \url{https://github.com/ropensci/gutenbergr/issues}
}

}
\author{
\strong{Maintainer}: Myfanwy Johnston \email{mrowlan1@gmail.com}

Authors:
\itemize{
  \item David Robinson \email{admiral.david@gmail.com} [copyright holder]
}

}
\keyword{internal}
