% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_mirrors.R
\name{is_working_gutenberg_mirror}
\alias{is_working_gutenberg_mirror}
\title{Check if a URL resolves to a working Gutenberg mirror}
\usage{
is_working_gutenberg_mirror(url)
}
\arguments{
\item{url}{An http(s) or ftp(s) URL to check.}
}
\value{
Boolean: whether the \code{url} resolves to a mirror.
}
\description{
Checks for a root level \code{README} file at \code{url} with reference to
\code{GUTINDEX.ALL}. If this exists, \code{url} is most likely a working
Gutenberg mirror.
}
\keyword{internal}
