% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_gutenberg_metadata.R
\name{load_cached_authors}
\alias{load_cached_authors}
\title{Load cached author metadata into memory}
\usage{
load_cached_authors(lib = getOption("gutenbergr_cache"))
}
\arguments{
\item{lib}{the directory containing cached metadata. Generated with \code{cache_gutenberg_metadata}}
}
\description{
Load cached author metadata into memory
}
