% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_gutenberg_metadata.R
\name{load_cached_formats}
\alias{load_cached_formats}
\title{Load cached available format metadata into memory}
\usage{
load_cached_formats(lib = getOption("gutenbergr_cache"), ret = FALSE)
}
\arguments{
\item{lib}{the directory containing cached metadata. Generated with \code{cache_gutenberg_metadata}}

\item{ret}{should the metadata data.frame be returned}
}
\description{
Load cached available format metadata into memory
}
\examples{
load_cached_formats()

load_cached_formats("~/gutenbergr")

options(gutenbergr_cache = "~/gutenbergr")
load_cached_formats()

gutenberg_formats <- load_cached_formats(ret = TRUE)
}
