% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_gutenberg_metadata.R
\name{load_cached_works}
\alias{load_cached_works}
\title{Load cached work metadata into memory}
\usage{
load_cached_works(lib = getOption("gutenbergr_cache"), ret = FALSE)
}
\arguments{
\item{lib}{the directory containing cached metadata. Generated with \code{cache_gutenberg_metadata}}

\item{ret}{should the metadata data.frame be returned}
}
\description{
Load cached work metadata into memory
}
\examples{
load_cached_works()

load_cached_works("~/gutenbergr")

options(gutenbergr_cache = "~/gutenbergr")
load_cached_works()

gutenberg_works <- load_cached_works(ret = TRUE)
}
