% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_gutenberg_metadata.R
\name{load_ed_works}
\alias{load_ed_works}
\title{Load cached work metadata into memory}
\usage{
load_ed_works(lib = getOption("gutenbergr_cache"))
}
\arguments{
\item{lib}{the directory containing cached metadata. Generated with \code{cache_gutenberg_metadata}}
}
\description{
Load cached work metadata into memory
}
\examples{
load_cached_works()

load_cached_works("~/gutenbergr")

options(gutenbergr_cache = "~/gutenbergr")
load_cached_works()
}
