% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{maybe_message}
\alias{maybe_message}
\title{Display a message conditionally}
\usage{
maybe_message(verbose, message, class = NULL, ..., call = rlang::caller_env())
}
\arguments{
\item{verbose}{Logical; whether to display the message.}

\item{message}{Message to display.}

\item{class}{Optional message class.}

\item{...}{Additional arguments passed to \code{cli::cli_inform()}.}

\item{call}{The execution environment of the calling function.}
}
\value{
Invisibly returns \code{NULL}. Called for its side effect of displaying a message.
}
\description{
Display a message conditionally
}
\keyword{internal}
