% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rdf_metadata.R
\name{parse_metadata_path}
\alias{parse_metadata_path}
\title{Parse and combine all metadata files in path}
\usage{
parse_metadata_path(path, pattern = "*.rdf", parser = parse_rdf_metadata)
}
\description{
Parse and combine all metadata files in path
}
