% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rdf_metadata.R
\name{parse_rdf_metadata}
\alias{parse_rdf_metadata}
\title{Parse a Gutenberg RDF/XML metadata file}
\usage{
parse_rdf_metadata(filepath)
}
\description{
/code{parse_rdf_metadata} takes a filepath and loads the file as an
XML file (assuming it is and RDF/XML format file. If the file is as
expected, the file is parsed to pull out metadata about the work, the
author, and the different file formats available.
}
