% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_download.R
\name{read_next}
\alias{read_next}
\title{Loop through paths to find a file}
\usage{
read_next(possible_urls)
}
\arguments{
\item{possible_urls}{URLs to try.}
}
\value{
A character vector of lines of text or NULL if the book could not be
downloaded.
}
\description{
Loop through paths to find a file
}
\keyword{internal}
