% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_url}
\alias{read_url}
\title{Read a file from a URL}
\usage{
read_url(url)
}
\arguments{
\item{url}{URL to a file}
}
\value{
A character vector of lines of text or \code{NULL} if the file could not be
downloaded.
}
\description{
Quietly download, read, and delete file
}
\keyword{internal}
