% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_books}
\alias{sample_books}
\title{Sample Book Downloads}
\format{
A tbl_df (from \code{\link[tibble:tibble]{tibble::tibble()}}) with one row for each
line of text from each book, with columns:
\describe{
\item{gutenberg_id}{Unique identifier for the work that can
be used to join with the \link{gutenberg_metadata} dataset.}
\item{text}{A character vector of lines of text.}
\item{title}{The title of this work.}
\item{author}{The author of this work.}
}
}
\usage{
sample_books
}
\description{
A tibble of book text for two sample books, generated using
\code{\link[=gutenberg_download]{gutenberg_download()}}.
}
\details{
This code was used to download the books:
\code{gutenberg_download(c(109, 105), meta_fields = c("title", "author"))}
}
\keyword{datasets}
