% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_gutenbergr_options}
\alias{set_gutenbergr_options}
\title{Set the package specific options for gutenbergr}
\usage{
set_gutenbergr_options(pkgname, gutenbergr_cache = "~/gutenbergr",
  gutenbergr_download_loc = tempdir())
}
\arguments{
\item{pkgname}{the name of the package being loaded}

\item{gutenbergr_cache}{the location to cache gutenbergr files such as metadata}
}
\description{
/code{set_gutenbergr_options} sets up the default options for the gutenbergr
package
}
\details{
This function is called when the package is loaded to define the options
that are used throughout the package. The options are initialised with
default values. The available options and their defaults are as follows:
}
