% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_download.R
\name{try_gutenberg_download}
\alias{try_gutenberg_download}
\title{Try to download book using various URLs}
\usage{
try_gutenberg_download(url)
}
\arguments{
\item{url}{The base URL of a book.}
}
\value{
A character vector of lines of text or NULL if the book could not be
downloaded.
}
\description{
Try to download book using various URLs
}
\keyword{internal}
