% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_gutenberg_metadata.R
\name{acquire_gutenberg_metadata}
\alias{acquire_gutenberg_metadata}
\title{Download and extract metadata}
\usage{
acquire_gutenberg_metadata(method = "curl",
  path = getOption("gutenbergr_download_loc"))
}
\arguments{
\item{method}{method passed to /code{download.file}}

\item{path}{location to store downloaded files}
}
\description{
/code{acquire_gutenberg_metadata} downloads (using curl as a default)
to download the metadata from Project Gutenberg. The compressed
file is stored in /code{tempdir()} and extracted there using
/code{untar}.
}
