% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_gutenberg_metadata.R
\name{cache_gutenberg_metadata}
\alias{cache_gutenberg_metadata}
\title{Acquire and cache metadata}
\usage{
cache_gutenberg_metadata(lib = getOption("gutenbergr_cache"),
  create_dir = TRUE)
}
\arguments{
\item{lib}{The location to store the cached metadata}

\item{create_dir}{Should the directory be created if it doesn't exist?}
}
\description{
/code{cache_gutenberg_metadata} calls /code{acquire_metadata}
to download and temporarily store the Project Gutenberg
metadata. It then formats and caches the metadata so
that it can be used again. Downloading and extracting
takes a long time, so it is useful to cache the data
}
\examples{
cache_gutenberg_metadata()

cache_gutenberg_metadata("~/gutenbergr")

options(gutenbergr_cache = "~/gutenbergr")
cache_gutenberg_metadata()
}
