% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discard_end_while}
\alias{discard_end_while}
\title{Discard all values at the start of .x while .p is true}
\usage{
discard_end_while(.x, .p)
}
\arguments{
\item{.x}{Vector}

\item{.p}{Logical vector}
}
\description{
Discard all values at the start of .x while .p is true
}
\keyword{internal}
