% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_cache.R
\name{gutenberg_cache_dir}
\alias{gutenberg_cache_dir}
\title{Get the active Gutenberg cache directory}
\usage{
gutenberg_cache_dir()
}
\value{
A character string representing the path to the cache directory.
}
\description{
Get the active Gutenberg cache directory
}
