% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_cache.R
\name{gutenberg_cache_path}
\alias{gutenberg_cache_path}
\title{Get the active cache directory path}
\usage{
gutenberg_cache_path()
}
\value{
A character string representing the path to the cache directory.
}
\description{
Calculates the path to the directory where Gutenberg files are stored,
based on the current \code{gutenbergr_cache_type} option.
}
\keyword{internal}
