% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_cache.R
\name{gutenberg_clear_cache}
\alias{gutenberg_clear_cache}
\title{Clear all files from the Gutenberg cache}
\usage{
gutenberg_clear_cache()
}
\value{
The number of files deleted (invisibly).
}
\description{
Deletes all cached \code{.rds} files in the directory currently returned by
\code{\link[=gutenberg_cache_dir]{gutenberg_cache_dir()}}.
}
