% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_download_books.R
\name{gutenberg_download}
\alias{gutenberg_download}
\title{Download a book using its Gutenberg ID}
\usage{
gutenberg_download(gutenberg_id, mirror = NULL)
}
\arguments{
\item{gutenberg_id}{A vector of Project Gutenberg ID}

\item{mirror}{Optionally a mirror URL to retrieve the books from}
}
\description{
Download a book using its Gutenberg ID
}

