% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_download.R
\name{gutenberg_get_mirror}
\alias{gutenberg_get_mirror}
\title{Get the recommended mirror for Gutenberg files}
\usage{
gutenberg_get_mirror(verbose = TRUE)
}
\arguments{
\item{verbose}{Whether to show messages about the Project Gutenberg mirror
that was chosen}
}
\value{
A character vector of the url for mirror to be used
}
\description{
Get the recommended mirror for Gutenberg files and set the global
\code{gutenberg_mirror} options.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

gutenberg_get_mirror()
\dontshow{\}) # examplesIf}
}
