% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_cache.R
\name{gutenberg_list_cache}
\alias{gutenberg_list_cache}
\title{List files in the Gutenberg cache}
\usage{
gutenberg_list_cache(verbose = TRUE)
}
\arguments{
\item{verbose}{Whether to show the status message showing the cache directory path.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with the following columns:
\describe{
\item{file}{The filename.}
\item{size_mb}{Size of the file in megabytes.}
\item{modified}{The last modification time.}
\item{path}{The file's absolute path.}
}
}
\description{
Provides a detailed list of files currently stored in the directory
returned by \code{\link[=gutenberg_cache_dir]{gutenberg_cache_dir()}}.
}
\keyword{cache}
