% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keep_while}
\alias{keep_while}
\title{Keep values at the start of .x while .p is true}
\usage{
keep_while(.x, .p)
}
\arguments{
\item{.x}{Vector}

\item{.p}{Logical vector}
}
\description{
Keep values at the start of .x while .p is true
}
\keyword{internal}
