% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_gutenberg_metadata.R
\name{load_cached_authors}
\alias{load_cached_authors}
\title{Load cached author metadata into memory}
\usage{
load_cached_authors(lib = getOption("gutenbergr_cache"), ret = FALSE)
}
\arguments{
\item{lib}{the directory containing cached metadata. Generated with \code{cache_gutenberg_metadata}}

\item{ret}{should the metadata data.frame be returned}
}
\description{
Load cached author metadata into memory
}
\examples{
load_cached_authors()

load_cached_authors("~/gutenbergr")

options(gutenbergr_cache = "~/gutenbergr")
load_cached_authors()

gutenberg_authors <- load_cached_authors(ret = TRUE)
}
