% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_url}
\alias{read_url}
\title{Read a file from a URL}
\usage{
read_url(url, ext = c(".zip", ".txt"))
}
\arguments{
\item{url}{URL to a file}

\item{ext}{Extension of the file to read}
}
\description{
Download, read, and delete file
}
\keyword{internal}
